requirements.txt
Django==2.0
pycryptodome==3.6.6
djangorestframework==3.7.7
django-cors-headers==2.1.0
requests==2.18.4
django-filter==1.1.0
psycopg2-binary==2.8.4
django-rest-swagger==2.2.0
coverage==4.5.2
django-nose==1.4.6
jsonnet==0.12.1
selenium
numpy
matplotlib
oauth2==1.9.0.post1
telegram
locust
django-bootstrap4
django-crispy-forms
python-telegram-bot